+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [1.0.1] - 2025-05-31
+
+### Network Manager and Improvements
+
+- Add a NetworkManager class to frequently check for internet reachability
+- Add a YieldTask utility class to yield a task inside a coroutine
+- Improve singletons
+
 ## [1.0.0] - 2025-05-07
 
 ### First Release